refactor: moves typo as separate file
