+++ packages/theme/src/theme.module.css
@@ -55,30 +55,6 @@
   --baseColor-other-ekspressen_500: #ED6C05;
   --baseColor-other-ekspressen_600: #C75B12;
 
-  --typo-heroTitle-fontSize: 2rem;
-  --typo-heroTitle-lineHeight: 2.5rem;
-
-  --typo-pageTitle-fontSize: 1.625rem;
-  --typo-pageTitle-lineHeight: 2rem;
-
-  --typo-sectionHeadline-fontSize: 1.4375rem;
-  --typo-sectionHeadline-lineHeight: 1.75rem;
-
-  --typo-itemHeadline-fontSize: 1.25rem;
-  --typo-itemHeadline-lineHeight: 1.5rem;
-
-  --typo-paragraphHeadline-fontSize: 1rem;
-  --typo-paragraphHeadline-lineHeight: 1.25rem;
-  --typo-paragraphHeadline-fontWeight: 600;
-
-  --typo-body-fontSize: 1rem;
-  --typo-body-lineHeight: 1.25rem;
-
-  --typo-lead-fontSize: 0.875rem;
-  --typo-lead-lineHeight: 1.25rem;
-
-  --typo-label-fontSize: 0.75rem;
-  --typo-label-lineHeight: 1rem;
 }
 
 /* Light theme data */
@@ -289,39 +265,4 @@
 .colors-transport_other {
   background-color: var(--colors-transport_other-backgroundColor);
   color: var(--colors-transport_other-color);
-}
-
-/* Typography definitions */
-.typo-heroTitle {
-  font-size: 2rem;
-  line-height: 2.5rem;
-}
-.typo-pageTitle {
-  font-size: 1.625rem;
-  line-height: 2rem;
-}
-.typo-sectionHeadline {
-  font-size: 1.4375rem;
-  line-height: 1.75rem;
-}
-.typo-itemHeadline {
-  font-size: 1.25rem;
-  line-height: 1.5rem;
-}
-.typo-paragraphHeadline {
-  font-size: 1rem;
-  line-height: 1.25rem;
-  font-weight: 600;
-}
-.typo-body {
-  font-size: 1rem;
-  line-height: 1.25rem;
-}
-.typo-lead {
-  font-size: 0.875rem;
-  line-height: 1.25rem;
-}
-.typo-label {
-  font-size: 0.75rem;
-  line-height: 1rem;
 }
